separated signup js file into login.js file - created utils folder withAuth - renamed signin to login for consistency - fixed links
