+++ controllers/homeRoutes.js
@@ -1,6 +1,8 @@
 const router = require("express").Router();
 const User = require("../models/User");
+const Subscription = require("../models/Subscription");
 const bcrypt = require("bcrypt");
+const withAuth = require("../utils/auth");
 
 router.get('/signup', (req, res) => {
   // If the user is already logged in, redirect the request to another route
@@ -11,6 +13,27 @@ router.get('/signup', (req, res) => {
   res.render('signup');
 });
 
+// Use withAuth middleware to prevent access to route
+router.get('/dashboard', withAuth, async (req, res) => {
+    try {
+      // Find the logged in user based on the session ID
+      const userData = await User.findByPk(req.session.user_id, {
+        attributes: { exclude: ['password'] },
+        include: [{ model: Subscription }],
+      });
+  
+      const user = userData.get({ plain: true });
+  
+      res.render('dashboard', {
+        ...user,
+        logged_in: true
+      });
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  });
+
+
 router.get('/login', (req, res) => {
   // If the user is already logged in, redirect the request to another route
   if (req.session.logged_in) {
@@ -18,8 +41,11 @@ router.get('/login', (req, res) => {
     return;
   }
 
-  res.render('signin');
+  res.render('login');
 });
+
+
+
 
 module.exports = router;
