+++ public/js/login.js
@@ -1,0 +1,24 @@
+const loginFormHandler = async (event) => {
+    event.preventDefault();
+  
+    const email = document.querySelector('#loginEmail').value.trim();
+    const password = document.querySelector('#loginPassword').value.trim();
+  
+    if (email && password) {
+      const response = await fetch('/api/users/login', {
+        method: 'POST',
+        body: JSON.stringify({ email, password }),
+        headers: { 'Content-Type': 'application/json' },
+      });
+  
+      if (response.ok) {
+        document.location.replace('/dashboard');
+      } else {
+        alert(response.statusText);
+      }
+    }
+  };
+
+  document
+  .querySelector('.login-form')
+  .addEventListener('submit', loginFormHandler);
