+++ public/js/signup.js
@@ -1,24 +1,3 @@
-const loginFormHandler = async (event) => {
-  event.preventDefault();
-
-  const email = document.querySelector('#loginEmail').value.trim();
-  const password = document.querySelector('#loginPassword').value.trim();
-
-  if (email && password) {
-    const response = await fetch('/api/users/login', {
-      method: 'POST',
-      body: JSON.stringify({ email, password }),
-      headers: { 'Content-Type': 'application/json' },
-    });
-
-    if (response.ok) {
-      document.location.replace('/dashboard');
-    } else {
-      alert(response.statusText);
-    }
-  }
-};
-
 const signupFormHandler = async (event) => {
   event.preventDefault();
 
@@ -40,10 +19,6 @@ const signupFormHandler = async (event) => {
     }
   }
 };
-
-document
-  .querySelector('.login-form')
-  .addEventListener('submit', loginFormHandler);
 
 document
   .querySelector('.signup-form')
