+++ views/login.handlebars
@@ -10,10 +10,10 @@
     <div class="container-fluid">
       <div class="row">
         <div class="col-7">
-          <h5 class="card-title"><a href="/">Sign Up</a></h5>
+          <h5 class="card-title"><a href="/signup">Sign Up</a></h5>
         </div>
         <div class="col-5">
-          <h5 class="card-title"><u>Sign In</u></h5>
+          <h5 class="card-title"><u>Log In</u></h5>
         </div>
       </div>
     </div>
@@ -27,7 +27,7 @@
       <input class="form-control" type="password" id="loginPassword" />
     </div>
     <div class="form-group">
-      <button class="btn" type="submit">Sign In</button>
+      <button class="btn" type="submit">Log In</button>
     </div>
   </form>
   </div>
@@ -36,4 +36,4 @@
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-<script src="./js/signup.js"></script>
+<script src="./js/login.js"></script>
