+++ views/signup.handlebars
@@ -13,7 +13,7 @@
           <h5 class="card-title"><u>Sign Up</u></h5>
         </div>
         <div class="col-5">
-          <h5 class="card-title"><a href="/signin">Sign In</a></h5>
+          <h5 class="card-title"><a href="/login">Log In</a></h5>
         </div>
       </div>
     </div>
